test: improve timezoneUtils tests for applyCustomFormat and error handling
